Limit dropna to named variables in ttest
rows are only dropped if they are nan for a named variable,
either res or xfac, in the ttest function. This ensures that rows
with data in res and xfac are not dropped simply because of nan in
other variables. This aligns output with ttest functions in scipy
and statsmodels.