change assert to warning in visuz.py
When any of significantly enriched genes, significantly depleted genes, or non-significant genes are missing, current implementation in visuz.py will throw an error.

However sometimes they could be missing. This is not biologically prohibited, and the code could handle this situation without problem.

So compared to an assertion, throwing out warning is more appropriate.